ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
python-multipart
